Rename bmi_dakota.py to bmi.py
It's in the dakota subpackage, so the 'dakota' in the name is
redundant.
